(EAI-988): Refactor `GenerateResponse` for tool call support  (#687)
* refactor GenerateRespose

* Clean up imports

* consolidate generate user prompt to the legacy file

* update test config imports

* Fix broken tests

* (EAI-989): Refactor verified answers to wrap `GenerateResponse` (#688)

verified answer generate response


* handle streaming

* separate generateresponse

* typo fix

---------
