+++ packages/mongodb-chatbot-server/src/processors/index.ts
@@ -1,9 +1,9 @@
 export * from "./FilterPreviousMessages";
-export * from "./GenerateUserPromptFunc";
 export * from "./MakeReferenceLinksFunc";
 export * from "./QueryPreprocessorFunc";
 export * from "./filterOnlySystemPrompt";
 export * from "./makeDefaultReferenceLinks";
 export * from "./makeFilterNPreviousMessages";
-export * from "./makeRagGenerateUserPrompt";
-export * from "./makeVerifiedAnswerGenerateUserPrompt";
+export * from "./makeVerifiedAnswerGenerateResponse";
+export * from "./includeChunksForMaxTokensPossible";
+export * from "./GenerateResponse";
