+++ packages/mongodb-chatbot-server/src/routes/conversations/addMessageToConversation.test.ts
@@ -94,7 +94,7 @@ describe("POST /conversations/:conversationId/messages", () => {
         .findOne({
           _id: new ObjectId(conversationId),
         });
-      expect(conversationInDb?.messages).toHaveLength(5); // system, user, assistant, user, assistant
+      expect(conversationInDb?.messages).toHaveLength(4); // user, assistant, user, assistant
     });
   });
 
@@ -331,97 +331,16 @@ describe("POST /conversations/:conversationId/messages", () => {
     });
   });
 
-  describe("Edge cases", () => {
-    test("Should respond with 200 and static response if query is negative toward MongoDB", async () => {
-      const query = REJECT_QUERY_CONTENT;
-      const res = await request(app)
-        .post(endpointUrl.replace(":conversationId", conversationId))
-        .set("X-FORWARDED-FOR", ipAddress)
-        .set("Origin", origin)
-        .send({ message: query });
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.content).toEqual(
-        defaultConversationConstants.NO_RELEVANT_CONTENT
-      );
-    });
-    test("Should respond with 200 and static response if no vector search content for user message", async () => {
-      const calledEndpoint = endpointUrl.replace(
-        ":conversationId",
-        conversationId
-      );
-      const response = await request(app)
-        .post(calledEndpoint)
-        .set("X-FORWARDED-FOR", ipAddress)
-        .set("Origin", origin)
-        .send({ message: NO_VECTOR_CONTENT });
-      expect(response.statusCode).toBe(200);
-      expect(response.body.references).toStrictEqual([]);
-      expect(response.body.content).toEqual(
-        defaultConversationConstants.NO_RELEVANT_CONTENT
-      );
-    });
-
-    describe("LLM not available but vector search is", () => {
-      const openAiClient = new OpenAI({
-        apiKey: "definitelyNotARealApiKey",
-      });
-      const brokenLlmService = makeOpenAiChatLlm({
-        openAiClient,
-        deployment: OPENAI_CHAT_COMPLETION_DEPLOYMENT,
-        openAiLmmConfigOptions: {
-          temperature: 0,
-          max_tokens: 500,
-        },
-      });
-
-      let conversationId: ObjectId,
-        conversations: ConversationsService,
-        app: Express;
-      let testMongo: Db;
-      beforeEach(async () => {
-        const { mongodb, appConfig } = await makeTestAppConfig();
-        testMongo = mongodb;
-        ({ app } = await makeTestApp({
-          ...appConfig,
-          conversationsRouterConfig: {
-            ...appConfig.conversationsRouterConfig,
-            llm: brokenLlmService,
-          },
-        }));
-
-        conversations = makeMongoDbConversationsService(testMongo);
-        const { _id } = await conversations.create({
-          initialMessages: [systemPrompt],
-        });
-        conversationId = _id;
-      });
-      test("should respond 200, static message, and vector search results", async () => {
-        const messageThatHasSearchResults = "Why use MongoDB?";
-        const response = await request(app)
-          .post(
-            endpointUrl.replace(":conversationId", conversationId.toString())
-          )
-          .set("X-FORWARDED-FOR", ipAddress)
-          .set("Origin", origin)
-          .send({ message: messageThatHasSearchResults });
-        expect(response.statusCode).toBe(200);
-        expect(response.body.content).toBe(
-          defaultConversationConstants.LLM_NOT_WORKING
-        );
-        expect(response.body.references.length).toBeGreaterThan(0);
-      });
-    });
-  });
-
   describe("create conversation with 'null' conversationId", () => {
     test("should create a new conversation with 'null' value for addMessageToConversation if configured", async () => {
+      const message = {
+        message: "hello",
+      };
       const res = await request(app)
         .post(DEFAULT_API_PREFIX + `/conversations/null/messages`)
         .set("X-FORWARDED-FOR", ipAddress)
         .set("Origin", origin)
-        .send({
-          message: "hello",
-        });
+        .send(message);
       expect(res.statusCode).toEqual(200);
       expect(res.body).toMatchObject({
         content: expect.any(String),
@@ -437,8 +356,12 @@ describe("POST /conversations/:conversationId/messages", () => {
       expect(conversation?._id.toString()).toEqual(
         res.body.metadata.conversationId
       );
-      expect(conversation?.messages).toHaveLength(3);
-      expect(conversation?.messages[0]).toMatchObject(systemPrompt);
+      expect(conversation?.messages).toHaveLength(2);
+      console.log(conversation?.messages[0]);
+      expect(conversation?.messages[0]).toMatchObject({
+        content: message.message,
+        role: "user",
+      });
     });
     test("should not create a new conversation with 'null' value for addMessageToConversation if NOT configured", async () => {
       const { app: appWithoutCustomData } = await makeTestApp({
